spotty/requirements.txt
bosdyn-api==3.1.2.1
bosdyn-choreography-client==3.1.2.1
bosdyn-choreography-protos==3.1.2.1
bosdyn-client==3.1.2.1
bosdyn-core==3.1.2.1
bosdyn-mission==3.1.2.1
faiss-cpu==1.8.0.post1
ffmpeg-python==0.2.0
grpcio==1.67.1
huggingface-hub==0.26.2
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
langsmith==0.1.147
llama_cpp_python==0.3.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
numpy==1.24.4
open3d==0.15.1
openai==1.58.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==24.2
Pillow==9.2.0
plyfile==0.7.4
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pvporcupine==3.0.3
pvrecorder==1.2.3
PyAudio==0.2.14
pycaw==20220416
pydantic==2.10.4
pygame==2.6.1
PyOpenGL==3.1.7
PyQt5==5.15.11
pyquaternion==0.9.9
python-dotenv==1.0.1
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.0.2
scipy==1.10.1
sentence-transformers==3.2.1
six==1.16.0
SQLAlchemy==2.0.36
sympy==1.13.3
tokenizers==0.20.3
tomlkit==0.11.4
torch==2.4.1
torchvision==0.19.1
tqdm==4.64.0
transformers==4.46.3
vtk==9.0.3
websocket-client==1.4.0
whispercpp @ git+https://github.com/stlukey/whispercpp.py@7af678159c29edb3bc2a51a72665073d58f2352f
zipp==3.8.1